OOP: add a check that random variables are != 0, and make the area method return int

+++ scr/main/java/figur/Figures.java
@@ -12,9 +12,7 @@ public abstract class Figures {
     }
 
     public abstract void coordinates();
-    public abstract double area();
-
-
+    public abstract int area();
 }
